# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: HCML Monitoring Portal (BFF)
  version: 1.0.0
  summary: A read-only backend-for-frontend over the Zabbix 7.0 JSON-RPC API.
  description: |
    The portal owns **no database**. Every response below is either passed through from Zabbix or
    derived in-process and cached in a 500-entry LRU that is lost on restart. Sites, links,
    incidents, host reachability and the monthly SLA have no table anywhere.

    **One write exists** in the whole surface: `POST /api/problems/acknowledge`. It uses a separate
    write-capable Zabbix token and is refused with 503 when that token is unset.

    ### Authentication

    Two schemes, and both work on every guarded path:

    - `Authorization: Bearer <jwt>`, the normal one, from `POST /api/auth/login`. 12-hour expiry.
    - `?token=<jwt>`: a query-string alternative. It exists because `EventSource` cannot set a
      header, but `auth.ts:129-132` reads it on **every** `/api/*` request, not only the SSE ones.
      `logging.ts` redacts it from the request log.

    When `AUTH_ENABLED=false` no route is guarded at all and every caller is treated as `admin`.

    ### Roles

    OpenAPI cannot express a role requirement per path, so each operation states its role in the
    description. From `ROUTE_RULES` (`server/src/auth.ts:35-40`), first match wins, unmatched is
    `viewer`:

    | Role | Paths |
    |---|---|
    | `admin` | `/api/reports/inventory` |
...
    | `viewer` | everything else |

    ### Errors

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - queryToken: []
paths:
...
  /api/problems:
    get:
      summary: Open problems
      description: |
        Cached 5 s, and **never served stale**: an operator who has just acknowledged must not be
        shown the previous list. About 664 problems at HCML, 459 kB compressing to 25 kB.
      operationId: getProblems
      tags: [Problems]
      responses:
        '200':
          description: Open problems, newest first.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Problem' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
...
  /api/sla:
    get:
      summary: Zabbix SLA definitions
      description: Cached 60 s.
      operationId: getSlas
      tags: [SLA]
      responses:
        '200':
          description: SLA objects.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ZbxSla' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ZabbixError' }
        '503': { $ref: '#/components/responses/ZabbixUnavailable' }
...
  /api/explain/problem:
    get:
      summary: Plain-language explanation of one problem
      description: |
        Cached **1 h**, keyed by the event id together with its acknowledged and resolved state, so
        an explanation is not reused after the problem changes.

        Buffered JSON, not streamed. This path sends the host name, trigger name, opdata, severity,
        start time, flags and the problem's tags to the model: in `hcml-portal` that means they
        leave the machine.
      operationId: explainProblem
      tags: [AI]
      parameters:
        - name: eventid
          in: query
          required: true
          schema: { type: string, pattern: '^\d+$' }
      responses:
        '200':
          description: The explanation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemExplanation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: '`not_found`, no open problem with that event id.'
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '502': { $ref: '#/components/responses/AiError' }
        '503': { $ref: '#/components/responses/AiUnavailable' }
  /api/explain/sla:
    get:
      summary: Plain-language explanation of an SLA figure
      description: |
        Cached 300 s. **Two mutually exclusive modes.**

        With `source=derived` it explains the portal's own monthly figure and takes `month`,
        `profile` and `scope`, where scope is `overall`, `site:<code>` or `category:<NAME>`; an
        unmatched scope is 404. Otherwise it explains a Zabbix-native SLA and requires `slaid`.

        `noData: true` marks an answer produced **without calling the model** because there was
        nothing to measure.
      operationId: explainSla
      tags: [AI]
      parameters:
        - name: source
          in: query
          description: '`derived` selects the portal''s figure. Any other value selects Zabbix''s.'
          schema: { type: string, enum: [derived] }
        - name: month
          in: query
          schema: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' }
        - name: profile
          in: query
          schema: { type: string, enum: [availability, hcml-report] }
        - name: scope
          in: query
          schema: { type: string, default: overall }
        - name: slaid
          in: query
          description: Required when `source` is not `derived`.
          schema: { type: string, pattern: '^\d+$' }
        - name: serviceid
          in: query
          schema: { type: string, pattern: '^\d+$' }
      responses:
        '200':
          description: The explanation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SlaAnswer' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: '`not_found`, unknown scope, or unknown `slaid`.'
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '502': { $ref: '#/components/responses/AiError' }
        '503': { $ref: '#/components/responses/AiUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The normal scheme. 12-hour expiry, carries `{ sub, role }`.
    queryToken:
      type: apiKey
      in: query
      name: token
      description: |
        The same JWT in the query string. Required for `EventSource`, which cannot set headers, but
        accepted on **every** guarded path. Kept out of the request log by `logging.ts`.
  responses:
    BadRequest:
      description: '`bad_request`, a parameter failed validation.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: '`unauthorized`, missing or expired portal JWT. Not the same as `zabbix_auth`.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    ZabbixError:
      description: '`zabbix_error`, unreachable, non-2xx, non-JSON-RPC, or a JSON-RPC error.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ZabbixUnavailable:
      description: |
        `zabbix_timeout` (no answer within `ZABBIX_TIMEOUT_MS`, default 10 s) or `zabbix_auth` (Zabbix
        rejected the API token). Both are configuration states rather than upstream faults, which is
        why they are 503 and not 504.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    AiError:
      description: '`ai_error`, the model answered, but not usably.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    AiUnavailable:
      description: |
        `ai_disabled` (not configured) or `ai_busy` (the single concurrency slot is taken: carries
        `Retry-After: 10`).
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [error, message]
      description: The only error shape this server produces, at every status.
      properties:
        error: { type: string, description: A stable machine-readable code. }
        message: { type: string, description: Safe to show a user. Never contains the Zabbix URL. }
...
    Tag:
      type: object
      required: [tag, value]
      properties:
        tag: { type: string }
        value: { type: string }
    Problem:
      type: object
      required: [eventid, objectid, object, name, severity, clock, acknowledged]
      properties:
        eventid: { type: string }
        objectid: { type: string }
        object: { type: string, description: '`0` is a trigger.' }
        name: { type: string }
        severity: { type: string, enum: ['0', '1', '2', '3', '4', '5'] }
        clock: { type: string }
        r_eventid: { type: string, description: '`0` while still open.' }
        r_clock: { type: string }
        acknowledged: { type: string, enum: ['0', '1'] }
        suppressed: { type: string, enum: ['0', '1'] }
        opdata: { type: string }
        tags: { type: array, items: { $ref: '#/components/schemas/Tag' } }
        host: { type: string, description: Enriched by the BFF. }
        hostid: { type: string, description: Enriched by the BFF. }
        manualClose:
          type: boolean
          description: Whether the trigger permits manual close. Gates the Close button.
      additionalProperties: true
...
    ZbxSla:
      type: object
      properties:
        slaid: { type: string }
        name: { type: string }
        period: { type: string }
        slo: { type: string }
        effective_date: { type: string }
        timezone: { type: string }
        status: { type: string }
      additionalProperties: true
...
    ProblemExplanation:
      type: object
      required: [summary, tagsExplained, businessImpact, recommendation]
      properties:
        summary: { type: string }
        tagsExplained:
          type: array
          items:
            type: object
            required: [tag, value, meaning]
            properties:
              tag: { type: string }
              value: { type: string }
              meaning: { type: string }
        businessImpact: { type: string }
        recommendation: { type: string }
    SlaAnswer:
      type: object
      required: [status, plain, meetingTarget, recommendation]
      properties:
        status: { type: string }
        plain: { type: string }
        meetingTarget: { type: boolean }
        recommendation: { type: string }
        noData:
          type: boolean
          description: True when there was nothing to measure, so no model call was made.